backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.3"
dependencies = [
    "anthropic>=0.84.0",
    "claude-agent-sdk>=0.1.45",
    "fastapi>=0.135.1",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.41.0",
]
